make sure everything is float32

+++ mothernet/datasets/__init__.py
@@ -89,8 +89,8 @@ def _encode_if_category(column: pd.Series | np.ndarray) -> pd.Series | np.ndarra
 def get_openml_classification(did, max_samples, multiclass=True, shuffled=True):
     dataset = openml.datasets.get_dataset(did, download_data=False, download_qualities=False, download_features_meta_data=False)
     X, y, categorical_indicator, attribute_names = dataset.get_data(target=dataset.default_target_attribute, dataset_format="dataframe")
-    X = np.array(X.apply(_encode_if_category))
-    y = np.array(_encode_if_category(y))
+    X = np.array(X.apply(_encode_if_category), dtype=np.float32)
+    y = np.array(_encode_if_category(y), dtype=np.float32)
     if not multiclass:
         X = X[y < 2]
         y = y[y < 2]
